Ecouter musique, lien liste artiste
Ajout d'une page pour écouter un morceau en particulier, ajout d'un lien pour chaque artiste dans la liste des artistes pour voir leurs musiques, ajout d'un lien dans la liste des musiques pour les écouter.

+++ src/main/java/fr/dawan/northCloud/dao/SongDao.java
@@ -30,7 +30,7 @@ public class SongDao {
 	}
 
 	@Transactional(readOnly = true)
-	public Song findById(int id) {
+	public Song findById(long id) {
 		return hibernateTemplate.get(Song.class, id);
 	}
